+++ test/Data/Store/StreamingSpec.hs
@@ -2,8 +2,8 @@
 {-# LANGUAGE OverloadedStrings #-}
 module Data.Store.StreamingSpec where
 
-import Control.Monad (void)
 import           Control.Exception (try)
+import           Control.Monad (void)
 import           Control.Monad.Trans.Resource
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Internal as BS
@@ -12,8 +12,8 @@ import qualified Data.Conduit.List as C
 import           Data.Int
 import           Data.List (unfoldr)
 import           Data.Monoid
-import           Data.Store.Impl (Poke (..))
-import           Data.Store.Internal -- (PeekException (..))
+import           Data.Store.Core.Internal (Poke(..))
+import           Data.Store.Internal
 import           Data.Store.Streaming
 import qualified System.IO.ByteBuffer as BB
 import           Test.Hspec